# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Nabra Space API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResult {
  error: string;
}

export interface DeleteResult {
  success: boolean;
}

export interface UserProfile {
  user_id: string;
  email: string;
  credits_balance: number;
  is_admin: boolean;
  is_banned: boolean;
  /** @nullable */
  full_name?: string | null;
  /** @nullable */
  avatar_url?: string | null;
  created_at: string;
}

export interface GenerationBlock {
  tone: string;
  text: string;
}

export interface Generation {
  id: string;
  user_id: string;
  /** @nullable */
  text_input?: string | null;
  /** @nullable */
  character_count?: number | null;
  /** @nullable */
  blocks_json?: GenerationBlock[] | null;
  /** @nullable */
  audio_url?: string | null;
  status: string;
  created_at: string;
}

export interface GenerationInput {
  blocks: GenerationBlock[];
}

export interface Transaction {
  id: string;
  user_id: string;
  amount_usd: number;
  /** @nullable */
  characters_granted?: number | null;
  /** @nullable */
  dodo_payment_id?: string | null;
  status: string;
  created_at: string;
}

export interface AffiliateStats {
  user_id: string;
  referral_code: string;
  commission_rate: number;
  total_clicks: number;
  total_signups: number;
  total_conversions: number;
  pending_balance: number;
  available_balance: number;
  status: string;
}

export interface PromoCode {
  id: string;
  code: string;
  characters_granted: number;
  max_total_uses: number;
  max_uses_per_user: number;
  current_uses: number;
  /** @nullable */
  expires_at?: string | null;
  /** @nullable */
  linked_affiliate_id?: string | null;
  is_active: boolean;
}

export interface PromoCodeInput {
  code: string;
  characters_granted: number;
  max_total_uses: number;
  max_uses_per_user?: number;
  /** @nullable */
  expires_at?: string | null;
  /** @nullable */
  linked_affiliate_id?: string | null;
}

export interface PromoCodeUpdate {
  is_active?: boolean;
  characters_granted?: number;
  max_total_uses?: number;
  /** @nullable */
  expires_at?: string | null;
}

export type CheckoutInputPackageId = typeof CheckoutInputPackageId[keyof typeof CheckoutInputPackageId];


export const CheckoutInputPackageId = {
  starter: 'starter',
  pro: 'pro',
  agency: 'agency',
} as const;

export interface CheckoutInput {
  package_id: CheckoutInputPackageId;
}

export interface CheckoutResult {
  url: string;
}

export interface AttributeInput {
  referral_code: string;
}

export interface AttributeResult {
  success: boolean;
  /** @nullable */
  message?: string | null;
}

export interface ClickInput {
  referral_code: string;
}

export interface ClickResult {
  success: boolean;
}

export interface RedeemInput {
  code: string;
}

export interface RedeemResult {
  success: boolean;
  characters_granted: number;
  new_balance: number;
  /** @nullable */
  message?: string | null;
}

export interface UserGrowthPoint {
  date: string;
  count: number;
}

export interface AdminTransaction {
  id: string;
  user_id: string;
  /** @nullable */
  email?: string | null;
  amount_usd: number;
  /** @nullable */
  characters_granted?: number | null;
  /** @nullable */
  dodo_payment_id?: string | null;
  status: string;
  /** @nullable */
  package_id?: string | null;
  created_at: string;
}

export interface AdminAffiliate {
  user_id: string;
  /** @nullable */
  email?: string | null;
  referral_code: string;
  commission_rate: number;
  total_clicks: number;
  total_signups: number;
  total_conversions: number;
  pending_balance: number;
  available_balance: number;
  status: string;
}

export interface AdminOverview {
  total_users: number;
  user_growth_pct: number;
  total_revenue: number;
  monthly_revenue: number;
  total_characters_generated: number;
  success_rate: number;
  recent_transactions?: AdminTransaction[];
  top_affiliates?: AdminAffiliate[];
  user_growth_data?: UserGrowthPoint[];
}

export interface AdminUser {
  user_id: string;
  email: string;
  /** @nullable */
  full_name?: string | null;
  /** @nullable */
  avatar_url?: string | null;
  created_at: string;
  credits_balance: number;
  generation_count: number;
  total_spent: number;
  is_admin: boolean;
  is_banned: boolean;
}

export type CreditsUpdateType = typeof CreditsUpdateType[keyof typeof CreditsUpdateType];


export const CreditsUpdateType = {
  set: 'set',
  add: 'add',
  subtract: 'subtract',
} as const;

export interface CreditsUpdate {
  type: CreditsUpdateType;
  amount: number;
}

export interface UserStatusUpdate {
  is_banned: boolean;
}

export interface AffiliateUpdate {
  commission_rate?: number;
  status?: string;
}

export interface PayoutInput {
  amount: number;
  /** @nullable */
  note?: string | null;
}

export interface PayoutRecord {
  id: string;
  affiliate_user_id?: string;
  amount: number;
  /** @nullable */
  note?: string | null;
  created_at: string;
}

export type ListAdminUsersParams = {
search?: string;
};

export type ListAdminTransactionsParams = {
status?: ListAdminTransactionsStatus;
};

export type ListAdminTransactionsStatus = typeof ListAdminTransactionsStatus[keyof typeof ListAdminTransactionsStatus];


export const ListAdminTransactionsStatus = {
  pending: 'pending',
  completed: 'completed',
  failed: 'failed',
} as const;