# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Nabra Space API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/overview:
    get:
      operationId: getAdminOverview
      tags: [admin]
      summary: Get admin overview stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Overview stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOverview"
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
  /admin/users/{id}/credits:
    patch:
      operationId: updateUserCredits
      tags: [admin]
      summary: Adjust user credits (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreditsUpdate"
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /admin/users/{id}/status:
    patch:
      operationId: updateUserStatus
      tags: [admin]
      summary: Ban or unban a user (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /admin/transactions:
    get:
      operationId: listAdminTransactions
      tags: [admin]
      summary: List all transactions (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, completed, failed]
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminTransaction"
  /admin/promo-codes:
    get:
      operationId: listAdminPromoCodes
      tags: [admin]
      summary: List all promo codes (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of promo codes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PromoCode"
    post:
      operationId: createAdminPromoCode
      tags: [admin]
      summary: Create promo code (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromoCodeInput"
      responses:
        "201":
          description: Created promo code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromoCode"
  /admin/promo-codes/{id}:
    patch:
      operationId: updateAdminPromoCode
      tags: [admin]
      summary: Update promo code (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromoCodeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromoCode"
    delete:
      operationId: deleteAdminPromoCode
      tags: [admin]
      summary: Delete promo code (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
  /admin/affiliates:
    get:
      operationId: listAdminAffiliates
      tags: [admin]
      summary: List all affiliates (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of affiliates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminAffiliate"
  /admin/affiliates/{id}:
    patch:
      operationId: updateAdminAffiliate
      tags: [admin]
      summary: Update affiliate commission/status (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AffiliateUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAffiliate"
  /admin/affiliates/{id}/payout:
    post:
      operationId: recordAffiliatePayout
      tags: [admin]
      summary: Record affiliate payout (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayoutInput"
      responses:
        "200":
          description: Payout recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayoutRecord"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    DeleteResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
    UserProfile:
      type: object
      required: [user_id, email, credits_balance, is_admin, created_at, is_banned]
      properties:
        user_id: { type: string }
        email: { type: string }
        credits_balance: { type: integer }
        is_admin: { type: boolean }
        is_banned: { type: boolean }
        full_name: { type: ["string", "null"] }
        avatar_url: { type: ["string", "null"] }
        created_at: { type: string }
...
    PromoCode:
      type: object
      required: [id, code, characters_granted, max_total_uses, max_uses_per_user, current_uses, is_active]
      properties:
        id: { type: string }
        code: { type: string }
        characters_granted: { type: integer }
        max_total_uses: { type: integer }
        max_uses_per_user: { type: integer }
        current_uses: { type: integer }
        expires_at: { type: ["string", "null"] }
        linked_affiliate_id: { type: ["string", "null"] }
        is_active: { type: boolean }
    PromoCodeInput:
      type: object
      required: [code, characters_granted, max_total_uses]
      properties:
        code: { type: string }
        characters_granted: { type: integer }
        max_total_uses: { type: integer }
        max_uses_per_user: { type: integer }
        expires_at: { type: ["string", "null"] }
        linked_affiliate_id: { type: ["string", "null"] }
    PromoCodeUpdate:
      type: object
      properties:
        is_active: { type: boolean }
        characters_granted: { type: integer }
        max_total_uses: { type: integer }
        expires_at: { type: ["string", "null"] }
...
    UserGrowthPoint:
      type: object
      required: [date, count]
      properties:
        date: { type: string }
        count: { type: integer }
    AdminOverview:
      type: object
      required: [total_users, user_growth_pct, total_revenue, monthly_revenue, total_characters_generated, success_rate]
      properties:
        total_users: { type: integer }
        user_growth_pct: { type: number }
        total_revenue: { type: number }
        monthly_revenue: { type: number }
        total_characters_generated: { type: integer }
        success_rate: { type: number }
        recent_transactions:
          type: array
          items:
            $ref: "#/components/schemas/AdminTransaction"
        top_affiliates:
          type: array
          items:
            $ref: "#/components/schemas/AdminAffiliate"
        user_growth_data:
          type: array
          items:
            $ref: "#/components/schemas/UserGrowthPoint"
    AdminUser:
      type: object
      required: [user_id, email, created_at, credits_balance, generation_count, total_spent, is_admin, is_banned]
      properties:
        user_id: { type: string }
        email: { type: string }
        full_name: { type: ["string", "null"] }
        avatar_url: { type: ["string", "null"] }
        created_at: { type: string }
        credits_balance: { type: integer }
        generation_count: { type: integer }
        total_spent: { type: number }
        is_admin: { type: boolean }
        is_banned: { type: boolean }
    AdminTransaction:
      type: object
      required: [id, user_id, amount_usd, status, created_at]
      properties:
        id: { type: string }
        user_id: { type: string }
        email: { type: ["string", "null"] }
        amount_usd: { type: number }
        characters_granted: { type: ["integer", "null"] }
        dodo_payment_id: { type: ["string", "null"] }
        status: { type: string }
        package_id: { type: ["string", "null"] }
        created_at: { type: string }
    AdminAffiliate:
      type: object
      required: [user_id, referral_code, commission_rate, total_clicks, total_signups, total_conversions, pending_balance, available_balance, status]
      properties:
        user_id: { type: string }
        email: { type: ["string", "null"] }
        referral_code: { type: string }
        commission_rate: { type: number }
        total_clicks: { type: integer }
        total_signups: { type: integer }
        total_conversions: { type: integer }
        pending_balance: { type: number }
        available_balance: { type: number }
        status: { type: string }
    CreditsUpdate:
      type: object
      required: [type, amount]
      properties:
        type:
          type: string
          enum: [set, add, subtract]
        amount: { type: integer }
    UserStatusUpdate:
      type: object
      required: [is_banned]
      properties:
        is_banned: { type: boolean }
    AffiliateUpdate:
      type: object
      properties:
        commission_rate: { type: number }
        status: { type: string }
    PayoutInput:
      type: object
      required: [amount]
      properties:
        amount: { type: number }
        note: { type: ["string", "null"] }
    PayoutRecord:
      type: object
      required: [id, amount, created_at]
      properties:
        id: { type: string }
        affiliate_user_id: { type: string }
        amount: { type: number }
        note: { type: ["string", "null"] }
        created_at: { type: string }