# lib/api-zod/src/generated/types/adminAffiliate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Nabra Space API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminAffiliate {
  user_id: string;
  /** @nullable */
  email?: string | null;
  referral_code: string;
  commission_rate: number;
  total_clicks: number;
  total_signups: number;
  total_conversions: number;
  pending_balance: number;
  available_balance: number;
  status: string;
}