# lib/api-zod/src/generated/types/payoutRecord.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Nabra Space API
 * OpenAPI spec version: 0.1.0
 */

export interface PayoutRecord {
  id: string;
  affiliate_user_id?: string;
  amount: number;
  /** @nullable */
  note?: string | null;
  created_at: string;
}