# lib/api-zod/src/generated/types/promoCode.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Nabra Space API
 * OpenAPI spec version: 0.1.0
 */

export interface PromoCode {
  id: string;
  code: string;
  characters_granted: number;
  max_total_uses: number;
  max_uses_per_user: number;
  current_uses: number;
  /** @nullable */
  expires_at?: string | null;
  /** @nullable */
  linked_affiliate_id?: string | null;
  is_active: boolean;
}