# lib/api-zod/src/generated/types/promoCodeUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Nabra Space API
 * OpenAPI spec version: 0.1.0
 */

export interface PromoCodeUpdate {
  is_active?: boolean;
  characters_granted?: number;
  max_total_uses?: number;
  /** @nullable */
  expires_at?: string | null;
}